requirements.txt
build==0.8.0
cycler==0.11.0
fonttools==4.37.1
kiwisolver==1.4.4
matplotlib==3.5.3
mpmath==1.2.1
numpy>=1.16.5,<=1.23.2
packaging==21.3
pandas>=1.3.5,<=1.4.4
pep517==0.13.0
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sympy>=1.10.1,<=1.11
tomli==2.0.1
pytest
flake8~=5.0.4
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
mysqlclient==2.1.1
python-dotenv==0.21.0
sqlparse==0.4.2
Django-MathJax